backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.6.1
pillow==10.0.0
opencv-python-headless==4.8.1.78
joblib==1.3.2
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0
tld==0.13.1
gunicorn==21.2.0